added a feign client configuration and reverted the application configuration
