+++ src/main/java/uk/gov/hmcts/probate/config/FeignClientConfiguration.java
@@ -1,0 +1,42 @@
+package uk.gov.hmcts.probate.config;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.apache.http.client.config.RequestConfig;
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.HttpClientBuilder;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+@ConfigurationProperties("resttemplate.httpclient")
+public class FeignClientConfiguration {
+
+    @Setter
+    @Getter
+    private int timeout;
+
+    @Bean
+    public RestTemplate restTemplate() {
+        RestTemplate restTemplate = new RestTemplate();
+        restTemplate.setRequestFactory(new HttpComponentsClientHttpRequestFactory(getHttpClient()));
+        return restTemplate;
+    }
+
+    private CloseableHttpClient getHttpClient() {
+        RequestConfig config = RequestConfig.custom()
+                .setConnectTimeout(timeout)
+                .setConnectionRequestTimeout(timeout)
+                .setSocketTimeout(timeout)
+                .build();
+
+        return HttpClientBuilder
+                .create()
+                .useSystemProperties()
+                .setDefaultRequestConfig(config)
+                .build();
+    }
+}
